refactor: move dtos to dto package
