+++ src/main/java/com/mandacarubroker/dtos/RequestLoginDTO.java
@@ -1,0 +1,6 @@
+package com.mandacarubroker.dtos;
+
+public record RequestLoginDTO (
+        String username,
+        String password
+){}
