+++ src/main/java/com/mandacarubroker/dtos/RequestStockDTO.java
@@ -1,4 +1,4 @@
-package com.mandacarubroker.domain.stock;
+package com.mandacarubroker.dtos;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
