+++ src/main/java/com/mandacarubroker/dtos/ResponseUserDTO.java
@@ -1,4 +1,6 @@
-package com.mandacarubroker.domain.user;
+package com.mandacarubroker.dtos;
+
+import com.mandacarubroker.domain.user.User;
 
 import java.time.LocalDate;
