+++ src/test/java/com/mandacarubroker/controller/StockControllerTest.java
@@ -2,7 +2,7 @@ package com.mandacarubroker.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mandacarubroker.controller.exceptions.StandardError;
-import com.mandacarubroker.domain.stock.RequestStockDTO;
+import com.mandacarubroker.dtos.RequestStockDTO;
 import com.mandacarubroker.domain.stock.Stock;
 import com.mandacarubroker.service.StockService;
 import jakarta.persistence.EntityNotFoundException;
